UnzipFiles done

+++ logfile/operations/types/unzip_files.py
@@ -5,12 +5,10 @@ Module contains file operation to unzip files
 import gzip
 import tarfile
 import logging
-from typing import Optional
 from pathlib import Path
 from logfile.operations.operation_base import OperationBase
 
 # pylint: disable=W1203
-# pylint: disable=no-else-return
 
 class UnzipFiles(OperationBase):
     """
@@ -23,7 +21,7 @@ class UnzipFiles(OperationBase):
             False takes current folder, true is recursive through current folder and subs
     """
     @staticmethod
-    def extract_tar(filepath: str) -> Optional[str]:
+    def extract_tar(filepath: str) -> str:
         """
         Extracting tar file
 
@@ -34,70 +32,64 @@ class UnzipFiles(OperationBase):
             str: Directory extracted in
             None: Tar extracting failed
         """
-        try:
+        if filepath and Path(filepath).exists():
             filepath = Path(filepath)
-
             extract_to = filepath.parent / filepath.stem
-            extract_to.mkdir()
 
             filepath = filepath.as_posix()
             extract_to = extract_to.as_posix()
-            logging.debug(f"Extracting '{filepath}' to '{extract_to}'")
 
-            tar = tarfile.open(filepath)
-            tar.extractall(extract_to)
-            tar.close()
+            try:
+                logging.debug(f"Create directory '{extract_to}'")
+                Path(extract_to).mkdir()
 
-            logging.debug("Deleting file '{filepath}'")
-            Path(filepath).unlink()
+                logging.debug(f"Extracting '{filepath}' to '{extract_to}'")
+                tar = tarfile.open(filepath)
+                tar.extractall(extract_to)
+                tar.close()
 
-            return extract_to
+                logging.debug("Deleting file '{filepath}'")
+                Path(filepath).unlink()
 
-        except OSError as exc:
-            logging.warning(f"Error while extracting file '{filepath}' {exc}")
+                return extract_to
+
+            except tarfile.ReadError as exc:
+                logging.warning(f"File is corrupted '{filepath}' {exc}")
+
+            except OSError as exc:
+                logging.warning(f"Could not delete '{filepath}' {exc}")
+
+        return None
 
     @staticmethod
-    def extract_gz(filepath: str) -> None:
+    def extract_gz(filepath: str) -> bool:
         """
         Extract gz file
 
         Args:
             filepath(str): Filepath to gz file
         """
-        try:
+        if filepath and Path(filepath).exists():
             filepath = Path(filepath)
-            extract_to = filepath.parent
+            extract_to = (filepath.parent / filepath.stem).as_posix()
             filepath = filepath.as_posix()
 
-            logging.debug(f"Extracting '{filepath}' to '{extract_to}'")
-            gzfile = gzip.open(filepath)
-            output = open(extract_to, "wb")
-            output.write(gzfile.read())
-            gzfile.close()
-            output.close()
+            try:
+                logging.debug(f"Extracting '{filepath}' to '{extract_to}'")
+                gzfile = gzip.open(filepath)
+                output = open(extract_to, "wb")
+                output.write(gzfile.read())
+                gzfile.close()
+                output.close()
 
-            Path(filepath).unlink()
-            logging.debug("Deleted file '{filepath}'")
+                logging.debug(f"Deleted file '{filepath}'")
+                Path(filepath).unlink()
+                return True
 
-        except OSError as exc:
-            logging.warning(f"Error while extracting file '{filepath}' {exc}")
+            except OSError as exc:
+                logging.warning(f"Error while extracting file '{filepath}' {exc}")
 
-    @staticmethod
-    def extract_file(filepath: str) -> Optional[str]:
-        """
-        Extracting tar or gz file
-
-        Args:
-            filepath(str): Filepath to compressed file
-        Returns:
-            str: Directory where files are extracted to if there is created a new directory
-        """
-        ext = Path(filepath).suffix.lower()
-        if ext in [".tgz", ".tar"]:
-            return UnzipFiles.extract_tar(filepath)
-        elif ext == ".gz":
-            UnzipFiles.extract_gz(filepath)
-        return None
+        return False
 
     @staticmethod
     def extract(filepath: str, recursive: bool) -> None:
@@ -108,10 +100,16 @@ class UnzipFiles(OperationBase):
             filepath(str): Filepath to compressed file
             Recursive(bool): Walk through tree and compress files inside
         """
-        new_dir = UnzipFiles.extract_file(filepath)
+        new_dir = None
+        if filepath:
+            ext = Path(filepath).suffix.lower()
+            if ext in [".tgz", ".tar"]:
+                new_dir = UnzipFiles.extract_tar(filepath)
+            elif ext == ".gz":
+                UnzipFiles.extract_gz(filepath)
 
         if recursive and new_dir:
-            filepaths = list(x for x in Path(new_dir).glob("**/*.*") if x.is_file())
+            filepaths = OperationBase.get_files(new_dir, True)
             for filep in filepaths:
                 UnzipFiles.extract(filep, recursive)
 
@@ -119,20 +117,14 @@ class UnzipFiles(OperationBase):
         """
         Extracting files inside directory with given instructions
         """
-        try:
-            directory = self._directory_instruction
-            recursive = self._recursive_instruction
+        directory = self.directory_instruction
+        recursive = self.recursive_instruction
 
-            # Dont enable recursive for this, as we want to only get files from current dir
-            files_in_directory = self._get_files(directory, False)
+        directory_path = self.make_directory_path(directory)
+        files_in_directory = self.get_files(directory_path, False)
+
+        if files_in_directory != []:
             for filepath in files_in_directory:
                 UnzipFiles.extract(filepath, recursive)
             return True
-
-        except TypeError as exc:
-            logging.warning(f"Instructions is '{self.instructions}' {exc}")
-
-        except OSError as exc:
-            logging.warning(f"Invalid relative path '{directory}' {exc}")
-
         return False
